Fix conflicts:
	desktop_version/src/Game.cpp
	desktop_version/src/Graphics.cpp
	desktop_version/src/Input.cpp
	desktop_version/src/Render.cpp
	desktop_version/src/UtilityClass.cpp
	desktop_version/src/editor.cpp
	desktop_version/src/main.cpp

+++ desktop_version/src/Finalclass.h
@@ -1,15 +1,12 @@
 #ifndef FINALCLASS_H
 #define FINALCLASS_H
 
-#include "Game.h"
-#include "Entity.h"
-
 #include <string>
 
 class finalclass
 {
 public:
-    const int* loadlevel(int rx, int ry);
+    const short* loadlevel(int rx, int ry);
 
     std::string roomname;
     bool warpx, warpy;
