+++ desktop_version/src/Labclass.h
@@ -1,15 +1,12 @@
 #ifndef LABCLASS_H
 #define LABCLASS_H
 
-#include "Game.h"
-#include "Entity.h"
-
 #include <string>
 
 class labclass
 {
 public:
-    const int* loadlevel(int rx, int ry);
+    const short* loadlevel(int rx, int ry);
 
     std::string roomname;
     int rcol;
