+++ desktop_version/src/WarpClass.h
@@ -1,15 +1,12 @@
 #ifndef WARPCLASS_H
 #define WARPCLASS_H
 
-#include "Game.h"
-#include "Entity.h"
-
 #include <string>
 
 class warpclass
 {
 public:
-	const int* loadlevel(int rx, int ry);
+	const short* loadlevel(int rx, int ry);
 	std::string roomname;
 	int rcol;
 	bool warpx, warpy;
